move add_gridlines from style to annotate
